refactor: create and use order service
